requirements.txt
MarkupSafe==2.1.5
pytest~=8.1.1
setuptools~=69.2.0
pytz~=2024.1
python-dotenv~=1.0.1